requirements.txt
google-api-python-client==1.6.2
playsound==1.2.1
PyAudio==0.2.9
PyYAML==3.11
SpeechRecognition==3.6.0
